pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skilltest"
version = "0.1.0"
description = "A portable, framework-agnostic testing framework for agent skills"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "httpx>=0.27.0",
]

[project.scripts]
skilltest = "skilltest.cli:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "pypdf>=4.0",
    "pdfplumber>=0.11",
    "Pillow>=10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git", ".tox", "dist", "build", "fixtures"]
